test: add tool calling and reasoning tests for frontend on GPT-OSS (#3636)
